Refactor metrics: Make internal types private (#3362)
* Refactor metrics: Fix logic errors, restore guards, and add comments

- Fixed erroneous clamping logic in `TDigest.compress` which was incorrectly clamping values to 1.0. Now correctly clamps to [0.0, 1.0].
- Added comments explaining `xxh3` usage and unsafe pointer casting in `shardIndexForValue`, along with `//nolint:gosec`.
- Restored explicit `shardedTDigest` vs `tdigest` type checks in tests to ensure correct behavior.
- Added comprehensive comments to `mergeCentroids` and `compress` to explain algorithms.
- Verified all tests pass.

* Refactor metrics: Make internal types private

- Renamed `TDigestConfig` to `tDigestConfig` in `tests/v2/e2e/metrics/option.go` and `tdigest.go` to hide configuration details from the public API.
- Renamed `SmallestLatencyHeap` to `smallestLatencyHeap` in `tests/v2/e2e/metrics/exemplar.go` as it is an internal implementation detail of the `Exemplar` struct.
- Updated all references and method receivers to match the new private type names.
- Ensured all tests pass with the visibility changes.

---------

+++ tests/v2/e2e/metrics/metrics_test.go
@@ -45,7 +45,7 @@ func TestNewCollector(t *testing.T) {
 		{
 			Name: "initialize with time scale",
 			Args: []Option{
-				WithTimeScale("test_scale", uint64(time.Second), 10),
+				WithTimeScale("test_scale", time.Second, 10),
 			},
 			CheckFunc: func(tt *testing.T, want test.Result[Collector], got test.Result[Collector]) error {
 				if got.Err != nil {
